Move node select related class to nodeSelection package
