+++ presto-main/src/main/java/com/facebook/presto/execution/scheduler/nodeSelection/NodeSelector.java
@@ -11,9 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.execution.scheduler;
+package com.facebook.presto.execution.scheduler.nodeSelection;
 
 import com.facebook.presto.execution.RemoteTask;
+import com.facebook.presto.execution.scheduler.BucketNodeMap;
+import com.facebook.presto.execution.scheduler.SplitPlacementResult;
 import com.facebook.presto.metadata.InternalNode;
 import com.facebook.presto.metadata.Split;
 import com.google.common.collect.ImmutableSet;
